Causifyxion API

Causifyxion — Module

Causifyxion is a library for creating causally-related variables, and sampling from them.

Module Index

Causifyxion.CausalVariable

Causifyxion.Possible

Causifyxion.ValueUnknownError

Causifyxion.causify

Causifyxion.dependson

Causifyxion.getvalue

Causifyxion.isknown

Causifyxion.isunknown

Causifyxion.nrand!

Causifyxion.rand!

Causifyxion.reset!

Causifyxion.resetandrand!

Causifyxion.setunknown!

Causifyxion.setvalue!

Causifyxion.@causify

API

Causifyxion.CausalVariable — Type

CausalVariable{T}

A mutable struct, and the main struct that this package provides. A CausalVariable is just a wrapper around a mechanism of sampling that variable. Create one using causify.

Example:

x = causify(Normal(0,1))

The point of this struct is to:

Keep track of dependencies between variables

Keep track of sampled values until they are resampled

This enables you to:

Sample from a few dependent variables conveniently without needing to wrap it in one ugly function / closure

Manually intervene in the system and resample with your interventions

Causifyxion.Possible — Type

Possible{T}

A sum type including Unknown as well as Known(x::T).

Causifyxion.ValueUnknownError — Type

ValueUnknownError

You tried to get the value of a CausalVariable that has not yet been sampled.

Causifyxion.causify — Method

causify(distr::Distribution)
causify(rand::Function, [::Type{T},] dependencies::CausalVariable...)

This is the main entrypoint (along with the more convenient @causify) to get a CausalVariable struct. There are a few options to call this function.

using Distributions

x = causify(Normal(0,1)) # first way 
y = causify(x) do x 
    x^2
end # second way, defines y to be x^2
z = causify(Float64, x, y) do x, y 
    x + y 
end # third way, defines z to be x + y, and type hints at z's eltype. 

You may find the type hinting useful in some scenarios. In our case, it is redundant and will automatically be inferred. Also see @causify for a more convenient way to call this function.

Causifyxion.dependson — Method

dependson(parent, child)

Checks where the parent depends on the child, recursively. If it doesn't depend, then obviously changing the child is not going to change parent.

Causifyxion.getvalue — Method

getvalue(causalvariable::CausalVariable)

Returns the value that is wrapped within the causalvariable.

Causifyxion.isknown — Method

isknown(causalvariable)

true is the causalvariable has a known value (i.e. has been previously sampled). false otherwise.

Causifyxion.isunknown — Method

isknown(causalvariable)

true is the causalvariable has a known value (i.e. has been previously sampled). false otherwise.

Causifyxion.nrand! — Method

nrand!([intervention_function!::Function, ]rv::CausalVariable...; n = 5)

Get n absolutely fresh samples of your variables. Super handy if you want to sample lots of stuff.

Examples:

x = causify(Uniform(0,1))
y = @causify x^2 # define y as x^2

mat = nrand!(x, y) # first column = 5 fresh samples of x, second column = 5 fresh samples of y; those samples are consistent in each row

@assert all(mat[:,1] .^ 2 .== mat[:,2]) # every value in column 1 squared equals every value in column 2

rigged_matrix = nrand!(x, y) do
    setvalue!(x, 0.5)
end 

@assert all(rigged_matrix[:,1] .== 0.5)
@assert all(rigged_matrix[:,2] .== 0.25)

# and then if you want...
using DataFrames
df = DataFrame(rigged_matrix, [:x, :y])

Causifyxion.rand! — Method

rand!(causalvariables...)

Samples each of the causalvariables supplied and returns the value of those answers. If the variable value is already known, it returns that known value.

Example:

x = causify(Uniform(0,1))
first_sample = rand!(x)
second_sample = rand!(x)

@assert first_sample == second_sample # note that rand! is not a completely fresh sample. It keeps known values around until you set the CausalVariable to `Unknown`. 

rand! recursively modifies all CausalVariables beneath it, by design, because it samples them all.

Causifyxion.reset! — Method

reset!(rv::CausalVariable...)

In short, this makes the CausalVariable forget its current sample so you can resample the CausalVariable. Makes each of the causalvariables supplied Unknown and recursively does this for all children. To do this non-recursively, rather use setunknown!.

Example:

x = causify(Uniform(0,1))
y = @causify x^2

first_sample = rand!(y)

@assert isknown(x) # This value is sampled through y
@assert isknown(y) # This value is sampled directly 

reset!(y)

@assert isunknown(x) # This value was reset through y
@assert isunknown(y) # This value was reset directly 

Causifyxion.resetandrand! — Method

resetandrand!([intervention_function::Function, ]rv::CausalVariable...)

Get an absolutely fresh sample of your variables.

Examples:

x = causify(Uniform(0,1))
y = @causify x^2
fresh_sample1 = resetandrand!(y)
fresh_sample2 = resetandrand!(y)

@assert fresh_sample1 != fresh_sample2 # these are 100% fresh samples

rigged_value = resetandrand!(y) do
    setvalue!(x, 0.5)
end 

@assert rigged_value == 0.25 

Causifyxion.setunknown! — Method

setunknown!(causalvariable)

Sets the value of the causal variable back to Unknown.

Causifyxion.setvalue! — Method

setvalue!(causalvariable::CausalVariable{T}, value::T)

Sets the value that is wrapped within the causalvariable. value must be of type T where T is the eltype of causalvariable.

Causifyxion.@causify — Macro

@causify([settings..., ] expr)

This wraps normal Julia expressions inside a causify() function, and is the most convenient way to create causal variables. This macro does not work for Distributions, for which you must use causify(). At the moment there may or may not be a heinous abomination of code making this work.

u = @causify Uniform(0,1) # ❌ usually not going to do what you want (unless you are wanted to parameterise the distribution itself with causal variables)
u = causify(Uniform(0,1)) # ✅ much better

The following rules apply:

1) You can causify expressions containing other CausalVariables

Example:

a = causify(Normal(1,1))
b = 20
x = @causify a + b
@assert x isa CausalVariable && x.dependson = (a,)

This is equivalent to doing:

x = causify(a) do a 
    a + b
end 

2) You can causify entire assignment expressions

Example:

a = causify(Normal(1,1))
b = 20
@causify x = a + b
@assert x isa CausalVariable && x.dependson = (a,)

3) You can causify entire begin ... end blocks

Example:

x = causify(Normal(0, 1))
y = @causify x^2

@causify begin 
    a = x + y 
    b = x + z 
    c = 15
    d = 15 + 19 
    e = d + a
end 

@assert e isa CausalVariable
@assert !(d isa CausalVariable)

4) You can pass settings to the macro

Example:

x = causify(Normal(0, 1))
y = @causify x^2

@causify begin 
    a = x + y 
    b = x + z 
    c = 15
    d = 15 + 19 
    e = d + a
end 

@assert e isa CausalVariable
@assert d isa CausalVariable && eltype(d) <: Int